CSP-2379 Relocated GetProvidersForLarsCodeQueryResultTests and GetProvidersFromLarsCodeRequestV2Tests to the Application.UnitTests.Courses.GetProvidersForLarsCode namespace
