+++ src/SFA.DAS.Roatp.Application.UnitTests/Courses/GetProvidersForLarsCode/GetProvidersFromLarsCodeRequestV2Tests.cs
@@ -6,7 +6,7 @@ using SFA.DAS.Roatp.Application.Courses.Queries.GetProvidersFromLarsCode.V2;
 using SFA.DAS.Roatp.Domain.Models;
 using SFA.DAS.Testing.AutoFixture;
 
-namespace SFA.DAS.Roatp.Api.UnitTests.Controllers.ExternalReadControllers.CoursesControllerTests;
+namespace SFA.DAS.Roatp.Application.UnitTests.Courses.GetProvidersForLarsCode;
 
 [TestFixture]
 public class GetProvidersFromLarsCodeRequestV2Tests
